dius: add support for sleeping using macos
